Removed EventReceiver and EventDispatcher
