+++ pkg/channel/references.go
@@ -18,6 +18,7 @@ package channel
 
 import (
 	"fmt"
+	"strings"
 )
 
 // ChannelReference references a Channel within the cluster by name and
@@ -29,4 +30,17 @@ type ChannelReference struct {
 
 func (r *ChannelReference) String() string {
 	return fmt.Sprintf("%s/%s", r.Namespace, r.Name)
+}
+
+// ParseChannel converts the channel's hostname into a channel
+// reference.
+func ParseChannel(host string) (ChannelReference, error) {
+	chunks := strings.Split(host, ".")
+	if len(chunks) < 2 {
+		return ChannelReference{}, fmt.Errorf("bad host format %q", host)
+	}
+	return ChannelReference{
+		Name:      chunks[0],
+		Namespace: chunks[1],
+	}, nil
 }
